+++ src/clients/inworld.client.ts
@@ -1,10 +1,6 @@
 import util = require('node:util');
 import { ServiceError } from '@grpc/grpc-js';
-import {
-  CapabilitiesRequest,
-  ClientRequest,
-  UserRequest,
-} from '@proto/world-engine_pb';
+import { CapabilitiesRequest, ClientRequest } from '@proto/world-engine_pb';
 
 import {
   ApiKey,
@@ -24,7 +20,7 @@ import { InworldConnectionService } from '../services/inworld_connection.service
 
 export class InworldClient {
   private apiKey: ApiKey | undefined;
-  private user: UserRequest;
+  private user: User;
   private scene: string = '';
   private client: ClientRequest;
   private config: InternalClientConfiguration;
@@ -49,7 +45,7 @@ export class InworldClient {
   }
 
   setUser(user: User) {
-    this.user = new UserRequest().setName(user.fullName);
+    this.user = user;
 
     return this;
   }